Added several categories. Added a class for WSSE. Added several front APIs.

+++ test/TestSimpleHttpClient.m
@@ -87,10 +87,10 @@
 {
     NSLog(@"assertCodeAndLengthWithContent: %@\n", context);
 
-    int status = [[_response objectForKey:context] statusCode];
+    NSInteger status = [[_response objectForKey:context] statusCode];
     NSAssert1(200 == status, @"status is %d.", status);
 
-    int length = [[_data objectForKey:context] length];
+    NSInteger length = [[_data objectForKey:context] length];
     NSAssert1(0 < length, @"%d byte.", length);
 
     NSLog(@"%d byte was received.\n", length);
